fix(joe-995,992,943): progressive port, façade peels, event-pump kernel
Single PR for recommended post-#964 Todo items to minimize CI cost:

- JOE-995: wire local WorkspaceSessionPort getSessionView + getSessionInfo;
  migrate ipc-runtime-context, session:diff, chart:save-artifact to the port
- JOE-992: peel mcp-formatters and work-store/work-queries under LOC budgets
- JOE-943: expand shared reconnect helpers; desktop + cloud use named kernel APIs
